requirements.txt
# numpy
numpy==2.0.2
 
# openai
openai-whisper
pydub 
ffmpeg-python